requirements.txt
langchain-openai>=0.3.30
langchain-core>=0.3.74
pandas>=2.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
openai>=1.99.9
langgraph>=0.2.45
typing-extensions>=4.12.0
xlsxwriter>=3.2.5
pydantic>=2.7.4
streamlit>=1.28.0
plotly>=5.17.0
python-dotenv>=1.0.0